# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  analysis.BatchView:
    properties:
      completed:
        type: integer
      createdAt:
        type: string
      done:
        type: boolean
      failed:
        type: integer
      id:
        example: batch_20260917_a1b2c3d4e5f6
        type: string
      percent:
        example: "42.50"
        type: string
      taskIds:
        items:
          type: string
        type: array
      total:
        type: integer
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  analysis.CancelView:
    properties:
      canceled:
        example: true
        type: boolean
    type: object
  analysis.ProgressView:
    properties:
      currentIndex:
        type: integer
      doneSteps:
        type: integer
      etaSeconds:
        type: integer
      message:
        type: string
      percent:
        example: "42.50"
        type: string
      steps:
        items:
          $ref: '#/definitions/value_objects.Step'
        type: array
      totalSteps:
        type: integer
      updatedAt:
        type: string
    type: object
  analysis.ResultView:
    properties:
      actionText:
        type: string
      createdAt:
        type: string
      decision:
        $ref: '#/definitions/value_objects.Decision'
      phases:
        items:
          $ref: '#/definitions/value_objects.PhaseOutcome'
        type: array
      reports:
        additionalProperties:
          type: string
        type: object
      symbol:
        example: 600519.SH
        type: string
      tradeDate:
        example: "2026-09-17"
        type: string
      usage:
        $ref: '#/definitions/value_objects.TokenUsage'
    type: object
  analysis.SubmitBatchRequest:
    properties:
      analysts:
        items:
          type: string
        type: array
      codes:
        example:
          - "600519"
          - "000001"
        items:
          type: string
        type: array
      depth:
        example: 3
        type: integer
      llmModel:
        type: string
      market:
        example: CN
        type: string
      tradeDate:
        example: "2026-09-17"
        type: string
    required:
      - codes
    type: object
  analysis.SubmitBatchView:
    properties:
      batch:
        $ref: '#/definitions/analysis.BatchView'
      tasks:
        items:
          $ref: '#/definitions/analysis.TaskView'
        type: array
    type: object
  analysis.SubmitRequest:
    properties:
      analysts:
        items:
          type: string
        type: array
      code:
        example: "600519"
        type: string
      depth:
        example: 3
        type: integer
      llmModel:
        type: string
      market:
        example: CN
        type: string
      tradeDate:
        example: "2026-09-17"
        type: string
    required:
      - code
    type: object
  analysis.TaskView:
    properties:
      analysts:
        items:
          type: string
        type: array
      attempts:
        type: integer
      batchId:
        type: string
      createdAt:
        type: string
      depth:
        example: 3
        type: integer
      error:
        type: string
      finishedAt:
        type: string
      id:
        example: task_20260917_a1b2c3d4e5f6
        type: string
      market:
        example: CN
        type: string
      progress:
        $ref: '#/definitions/analysis.ProgressView'
      startedAt:
        type: string
      status:
        example: running
        type: string
      statusText:
        type: string
      symbol:
        example: 600519.SH
        type: string
      tradeDate:
        example: "2026-09-17"
        type: string
      userId:
        type: integer
    type: object
...
  response.Envelope:
    properties:
      code:
        type: integer
      data: { }
      message:
        type: string
    type: object
  response.PageData:
    properties:
      items: { }
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  value_objects.Action:
    enum:
      - buy
      - hold
      - sell
      - reduce
      - increase
      - undecided
    type: string
    x-enum-varnames:
      - ActionBuy
      - ActionHold
      - ActionSell
      - ActionReduce
      - ActionIncrease
      - ActionUndecided
  value_objects.Decision:
    properties:
      action:
        $ref: '#/definitions/value_objects.Action'
      confidence:
        description: 0-1
        type: number
      position:
        description: 建议仓位百分比 0-100
        type: number
      reasoning:
        type: string
      risk_score:
        description: 0-10，越高越危险
        type: number
      stop_loss:
        type: number
      summary:
        type: string
      target_price:
        type: number
    type: object
...
  value_objects.PhaseOutcome:
    properties:
      agents:
        items:
          type: string
        type: array
      duration_seconds:
        type: number
      failed:
        items:
          type: string
        type: array
      phase:
        type: string
    type: object
...
  value_objects.Step:
    properties:
      detail:
        type: string
      done:
        type: boolean
      elapsed_seconds:
        type: integer
      failed:
        type: boolean
      key:
        type: string
      name:
        type: string
    type: object
  value_objects.TokenUsage:
    properties:
      calls:
        type: integer
      completion_tokens:
        type: integer
      cost_usd:
        type: number
      prompt_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
...
info:
  contact: { }
  description: |-
    多智能体股票分析平台 HTTP 接口。

    所有响应共用同一个信封 {"code":0,"message":"ok","data":{...}}，
    code 为 0 表示成功，非 0 时 data 恒为 null，具体取值见 response 包。

    健康检查 GET /healthz 不在 /api/v1 之下，因此未收录于本文档。

    免责声明：本平台仅用于学习与研究，不构成投资建议。分析结论由大模型生成，
    可能包含事实性错误。过往表现不代表未来收益，投资有风险，可能损失本金。
  title: Trading Agents API
  version: "1.0"
paths:
...
  /analysis/batches:
    post:
      consumes:
        - application/json
      parameters:
        - description: 批量分析请求，单批次股票数量有上限
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/analysis.SubmitBatchRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/analysis.SubmitBatchView'
                type: object
        "400":
          description: 请求参数不合法或超出单批次上限
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "429":
          description: 并发分析数已达上限
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: 批次落库或入队失败
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 批量提交分析任务
      tags:
        - 分析任务
  /analysis/batches/{id}:
    get:
      parameters:
        - description: 批次 ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/analysis.BatchView'
                type: object
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 批次不存在
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 查询分析批次
      tags:
        - 分析任务
  /analysis/tasks:
    get:
      parameters:
        - description: 目标用户 ID，仅管理员可传，缺省为自己
          in: query
          name: userId
          type: integer
        - description: 任务状态过滤，留空表示不限
          enum:
            - queued
            - running
            - completed
            - failed
            - canceled
          in: query
          name: status
          type: string
        - description: 页码，默认 1
          in: query
          name: page
          type: integer
        - description: 每页条数，默认 20
          in: query
          name: pageSize
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/response.PageData'
                      - properties:
                          items:
                            items:
                              $ref: '#/definitions/analysis.TaskView'
                            type: array
                        type: object
                type: object
        "400":
          description: 状态取值非法
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: 无权查看其他用户的分析任务
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 分析任务列表
      tags:
        - 分析任务
    post:
      consumes:
        - application/json
      parameters:
        - description: 分析请求，depth 取 1/3/5 分别对应快速、标准、深度
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/analysis.SubmitRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/analysis.TaskView'
                type: object
        "400":
          description: 请求参数不合法
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "429":
          description: 并发分析数已达上限
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: 任务落库或入队失败
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 提交分析任务
      tags:
        - 分析任务
  /analysis/tasks/{id}:
    get:
      parameters:
        - description: 任务 ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/analysis.TaskView'
                type: object
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 查询分析任务
      tags:
        - 分析任务
  /analysis/tasks/{id}/cancel:
    post:
      parameters:
        - description: 任务 ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/analysis.CancelView'
                type: object
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: 任务已处于终态或被并发改写
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 取消分析任务
      tags:
        - 分析任务
  /analysis/tasks/{id}/progress:
    get:
      description: |-
        以 Server-Sent Events 推送任务进度，直到任务终结或客户端断开。
        正常帧为 `event: progress`，data 是统一信封且 data 字段为 progressView；
        出错帧为 `event: error`，信封的 data 为 null。
        长时间没有新进度时服务端发一行 `: keep-alive` 注释保活，客户端应忽略。
      parameters:
        - description: 任务 ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - text/event-stream
      responses:
        "200":
          description: 'SSE 事件流，每条形如 data: {...}\n\n'
          schema:
            type: string
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: 订阅进度流失败
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 订阅任务进度（SSE）
      tags:
        - 分析任务
  /analysis/tasks/{id}/result:
    get:
      parameters:
        - description: 任务 ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Envelope'
              - properties:
                  data:
                    $ref: '#/definitions/analysis.ResultView'
                type: object
        "401":
          description: 未登录或令牌无效
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: 任务不存在或尚未产出结果
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
        - BearerAuth: [ ]
      summary: 查询分析结果
      tags:
        - 分析任务
...
securityDefinitions:
  BearerAuth:
    description: 填入 "Bearer {accessToken}"，令牌由 POST /auth/login 获取。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...